Don’t run tests twice in custom binary test class
Moving the helper functions to the base test class we can now extend that one intead. Otherwise, all tests would have been executed twice.
